# server/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"IOrderResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"tableId": {
						"type": "string"
					},
					"statusId": {
						"type": "string"
					}
				},
				"required": [
					"tableId",
					"statusId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "express-typescript",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/orders/search": {
			"post": {
				"operationId": "Search",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IOrderResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"OrdersService"
				],
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": []
			}
		},
		"/orders/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrderResponse"
								}
							}
						}
					}
				},
				"tags": [
					"OrdersService"
				],
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IOrderResponse"
							}
						}
					}
				}
			}
		},
		"/orders/update": {
			"post": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IOrderResponse"
								}
							}
						}
					}
				},
				"tags": [
					"OrdersService"
				],
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IOrderResponse"
							}
						}
					}
				}
			}
		},
		"/orders/remove": {
			"post": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"OrdersService"
				],
				"security": [
					{
						"Bearer": [
							"AuthService"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}